Miscellaneous Changes
A score counter symbol has been implemented. Coin counter has been fixed and a visual output has been included (doesnt work right tho). Barry now dies upon a hit (finally!!!)

+++ game.hpp
@@ -35,7 +35,7 @@ class Game{
 public:
     Mix_Music *bgMusic = NULL;
     Mix_Music *jetpacksound = NULL;
-    int bgspeed;
+    float bgspeed;
     bool init();
     bool loadMedia();
     void close();
